New hook logic on council module
